requirements.txt
openai>=1.60.0
moviepy>=1.0.3
python-dotenv>=1.0.0
requests>=2.31.0
google-genai>=0.1.0


